sfos: zypp Backend: Look for patterns matching distroId and install
When doing a dist-upgrade, the distroId can be the name of a pattern
(without the "pattern:" prefix) to be marked for installation. If a
pattern with that name is not found, the distroId is ignored and the
dist-upgrade will be carried out as normal.

For the pattern installation to take place, the type of dist-upgrade
must be "complete". If the type of dist-upgrade is "minimal", and the
pattern exists, the pattern's dependencies will be downloaded, but
not installed (as "minimal" signifies "download only" in our case).

backends: zypp: Install all patterns matching distroId on upgrade