sfos: Add libzypp authentication report handler
To avoid infinite loops in the authentication phase for repos for
which no authentication exists, connect an authentication handler
that logs the authentication request and then fails (as we do not
have or want interactive authentication in the zypp backend).

backends: zypp: Handle authentication reports by failing immediately

Squashed:

zypp: Force failure for missing ssu.ini

In cases of a missing ssu.ini, the do { ... } while (retry) loop in
zypp/media/MediaCurl.cc does not exit, even if MediaCurl::authenticate()
(and in turn our implementation of prompt()) returns false (=abort).

To fix this, we now throw a libzypp zypp::media::MediaException in the
prompt() function in order to force the backend and libzypp to give up.