[backends] zypp: Force failure for missing ssu.ini
In cases of a missing ssu.ini, the do { ... } while (retry) loop in
zypp/media/MediaCurl.cc does not exit, even if MediaCurl::authenticate()
(and in turn our implementation of prompt()) returns false (=abort).

To fix this, we now throw a libzypp zypp::media::MediaException in the
prompt() function in order to force the backend and libzypp to give up.