[backends] zypp: Report progress for download and install separately
This avoids situations where in case all packages were already
downloaded, the overall progress would only go to 50%. Ideally, we'd get
info on whether or not a package was already downloaded from libzypp,
but we don't have a recent enough version in it yet. Once that's
available, we can fully fix the percentage reporting for downloads.