[zypp] Honor allow_deps argument. JB#51760
Given a package that is required by other package(s), the allow_deps
argument is meant to control whether it may remove the package and
everything requiring it or fail.

The argument to zypp_perform_execution that was originally hardcoded as
TRUE is forwarded to zypp::Resolver::setForceResolve.

Handling allow_deps this way, although not ideal, is sufficient to fix
our use case and shouldn't make things worse.