sfos: Refresh the repo cache before resolving. Fixes JB#52744
On resolve, check that each enabled repo is cached, and refresh them all
if there's one that isn't. Packages cannot resolve if the repo caches
are missing, which means that system packages will be flagged as
problematic during upgrade.

If developer mode is being installed offline, this refresh will fail so
the resolve should continue anyway. But at least we tried.